requirements.txt
numpy==1.22.3
pandas==1.4.2
streamlit==1.9.0
opencv-python-headless==4.5.3.56
Pillow~=9.1.0
scikit-learn~=1.0.2